Regorganised project fixes #1
