+++ client/src/App.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { Switch, Route, BrowserRouter } from 'react-router-dom';
 
-import Home from './components/Home';
-import Signup from './components/Signup';
-import Signin from './components/Signin';
-import HomeDash from './components/HomeDash';
+import Home from './components/layouts/Home';
+import Signup from './components/containers/Signup';
+import Signin from './components/containers/Signin';
+import HomeDash from './components/layouts/HomeDash';
 
 function AppTest() {
   return (
