+++ client/src/components/containers/HomeDash.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import HomeDashLayout from '../layouts/HomeDash';
+
+export default function () {
+  return (
+    <>
+      <HomeDashLayout />
+    </>
+  );
+}
