+++ client/src/components/containers/Signin.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import SigninLayout from '../layouts/Signin';
+
+export default function () {
+  return (
+    <>
+      <SigninLayout />
+    </>
+  );
+}
