+++ client/src/components/containers/Signup.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import SignupLayout from '../layouts/Signup';
+
+export default function () {
+  return (
+    <>
+      <SignupLayout />
+    </>
+  );
+}
