+++ client/src/components/layouts/Home.jsx
@@ -19,9 +19,9 @@ import CssBaseline from '@material-ui/core/CssBaseline';
 import { makeStyles } from '@material-ui/core/styles';
 
 import Copyright from './Copyright';
-import HomeBGOne from '../images/Home_1.svg';
-import HomeBGTwo from '../images/Home_2.svg';
-import HomeBGThree from '../images/Home_3.svg';
+import HomeBGOne from '../../images/Home_1.svg';
+import HomeBGTwo from '../../images/Home_2.svg';
+import HomeBGThree from '../../images/Home_3.svg';
 import AuthHeader from './Auth';
 
 const useStyles = makeStyles((theme) => ({
