+++ client/src/components/layouts/Signin.jsx
@@ -2,6 +2,8 @@ import React from 'react';
 import Button from '@material-ui/core/Button';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import TextField from '@material-ui/core/TextField';
+import FormControlLabel from '@material-ui/core/FormControlLabel';
+import Checkbox from '@material-ui/core/Checkbox';
 import Link from '@material-ui/core/Link';
 import Grid from '@material-ui/core/Grid';
 import Box from '@material-ui/core/Box';
@@ -12,7 +14,7 @@ import Paper from '@material-ui/core/Paper';
 
 import AuthHeader from './Auth';
 import Copyright from './Copyright';
-import SignupBG from '../images/Signup.svg';
+import SigninBG from '../../images/Signin.svg';
 
 const useStyles = makeStyles((theme) => ({
   paper: {
@@ -22,15 +24,19 @@ const useStyles = makeStyles((theme) => ({
     alignItems: 'center',
     padding: theme.spacing(3),
   },
+  avatar: {
+    margin: theme.spacing(1),
+    backgroundColor: theme.palette.secondary.main,
+  },
   form: {
     width: '100%', // Fix IE 11 issue.
-    marginTop: theme.spacing(3),
+    marginTop: theme.spacing(1),
   },
   submit: {
     margin: theme.spacing(3, 0, 2),
   },
   backdrop: {
-    backgroundImage: `url(${SignupBG})`,
+    backgroundImage: `url(${SigninBG})`,
   },
   setBackground: {
     backgroundRepeat: 'no-repeat',
@@ -41,7 +47,7 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-export default function SignUp() {
+export default function SignIn() {
   const classes = useStyles();
 
   return (
@@ -51,57 +57,35 @@ export default function SignUp() {
           <CssBaseline />
           <Paper className={classes.paper}>
             <Typography component="h1" variant="h5">
-              Sign up
+              Sign in
             </Typography>
             <form className={classes.form} noValidate>
-              <Grid container spacing={2}>
-                <Grid item xs={12} sm={6}>
-                  <TextField
-                    autoComplete="fname"
-                    name="firstName"
-                    variant="outlined"
-                    required
-                    fullWidth
-                    id="firstName"
-                    label="First Name"
-                    autoFocus
-                  />
-                </Grid>
-                <Grid item xs={12} sm={6}>
-                  <TextField
-                    variant="outlined"
-                    required
-                    fullWidth
-                    id="lastName"
-                    label="Last Name"
-                    name="lastName"
-                    autoComplete="lname"
-                  />
-                </Grid>
-                <Grid item xs={12}>
-                  <TextField
-                    variant="outlined"
-                    required
-                    fullWidth
-                    id="email"
-                    label="Email Address"
-                    name="email"
-                    autoComplete="email"
-                  />
-                </Grid>
-                <Grid item xs={12}>
-                  <TextField
-                    variant="outlined"
-                    required
-                    fullWidth
-                    name="password"
-                    label="Password"
-                    type="password"
-                    id="password"
-                    autoComplete="current-password"
-                  />
-                </Grid>
-              </Grid>
+              <TextField
+                variant="outlined"
+                margin="normal"
+                required
+                fullWidth
+                id="email"
+                label="Email Address"
+                name="email"
+                autoComplete="email"
+                autoFocus
+              />
+              <TextField
+                variant="outlined"
+                margin="normal"
+                required
+                fullWidth
+                name="password"
+                label="Password"
+                type="password"
+                id="password"
+                autoComplete="current-password"
+              />
+              <FormControlLabel
+                control={<Checkbox value="remember" color="primary" />}
+                label="Remember me"
+              />
               <Button
                 type="submit"
                 fullWidth
@@ -109,18 +93,23 @@ export default function SignUp() {
                 color="primary"
                 className={classes.submit}
               >
-                Sign Up
+                Sign In
               </Button>
-              <Grid container justify="flex-end">
+              <Grid container>
+                <Grid item xs>
+                  <Link href="/password-reset" variant="body2">
+                    Forgot password?
+                  </Link>
+                </Grid>
                 <Grid item>
-                  <Link href="/signin" variant="body2">
-                    Already have an account? Sign in
+                  <Link href="/signup" variant="body2">
+                    Do not have an account? Sign Up
                   </Link>
                 </Grid>
               </Grid>
             </form>
           </Paper>
-          <Box mt={5}>
+          <Box mt={8}>
             <Copyright yourWebsite="My Diary App Demo" />
           </Box>
         </Container>
