+++ client/src/components/layouts/Signup.jsx
@@ -2,8 +2,6 @@ import React from 'react';
 import Button from '@material-ui/core/Button';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import TextField from '@material-ui/core/TextField';
-import FormControlLabel from '@material-ui/core/FormControlLabel';
-import Checkbox from '@material-ui/core/Checkbox';
 import Link from '@material-ui/core/Link';
 import Grid from '@material-ui/core/Grid';
 import Box from '@material-ui/core/Box';
@@ -14,7 +12,7 @@ import Paper from '@material-ui/core/Paper';
 
 import AuthHeader from './Auth';
 import Copyright from './Copyright';
-import SigninBG from '../images/Signin.svg';
+import SignupBG from '../../images/Signup.svg';
 
 const useStyles = makeStyles((theme) => ({
   paper: {
@@ -24,19 +22,15 @@ const useStyles = makeStyles((theme) => ({
     alignItems: 'center',
     padding: theme.spacing(3),
   },
-  avatar: {
-    margin: theme.spacing(1),
-    backgroundColor: theme.palette.secondary.main,
-  },
   form: {
     width: '100%', // Fix IE 11 issue.
-    marginTop: theme.spacing(1),
+    marginTop: theme.spacing(3),
   },
   submit: {
     margin: theme.spacing(3, 0, 2),
   },
   backdrop: {
-    backgroundImage: `url(${SigninBG})`,
+    backgroundImage: `url(${SignupBG})`,
   },
   setBackground: {
     backgroundRepeat: 'no-repeat',
@@ -47,7 +41,7 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-export default function SignIn() {
+export default function SignUp() {
   const classes = useStyles();
 
   return (
@@ -57,35 +51,57 @@ export default function SignIn() {
           <CssBaseline />
           <Paper className={classes.paper}>
             <Typography component="h1" variant="h5">
-              Sign in
+              Sign up
             </Typography>
             <form className={classes.form} noValidate>
-              <TextField
-                variant="outlined"
-                margin="normal"
-                required
-                fullWidth
-                id="email"
-                label="Email Address"
-                name="email"
-                autoComplete="email"
-                autoFocus
-              />
-              <TextField
-                variant="outlined"
-                margin="normal"
-                required
-                fullWidth
-                name="password"
-                label="Password"
-                type="password"
-                id="password"
-                autoComplete="current-password"
-              />
-              <FormControlLabel
-                control={<Checkbox value="remember" color="primary" />}
-                label="Remember me"
-              />
+              <Grid container spacing={2}>
+                <Grid item xs={12} sm={6}>
+                  <TextField
+                    autoComplete="fname"
+                    name="firstName"
+                    variant="outlined"
+                    required
+                    fullWidth
+                    id="firstName"
+                    label="First Name"
+                    autoFocus
+                  />
+                </Grid>
+                <Grid item xs={12} sm={6}>
+                  <TextField
+                    variant="outlined"
+                    required
+                    fullWidth
+                    id="lastName"
+                    label="Last Name"
+                    name="lastName"
+                    autoComplete="lname"
+                  />
+                </Grid>
+                <Grid item xs={12}>
+                  <TextField
+                    variant="outlined"
+                    required
+                    fullWidth
+                    id="email"
+                    label="Email Address"
+                    name="email"
+                    autoComplete="email"
+                  />
+                </Grid>
+                <Grid item xs={12}>
+                  <TextField
+                    variant="outlined"
+                    required
+                    fullWidth
+                    name="password"
+                    label="Password"
+                    type="password"
+                    id="password"
+                    autoComplete="current-password"
+                  />
+                </Grid>
+              </Grid>
               <Button
                 type="submit"
                 fullWidth
@@ -93,23 +109,18 @@ export default function SignIn() {
                 color="primary"
                 className={classes.submit}
               >
-                Sign In
+                Sign Up
               </Button>
-              <Grid container>
-                <Grid item xs>
-                  <Link href="/password-reset" variant="body2">
-                    Forgot password?
-                  </Link>
-                </Grid>
+              <Grid container justify="flex-end">
                 <Grid item>
-                  <Link href="/signup" variant="body2">
-                    Do not have an account? Sign Up
+                  <Link href="/signin" variant="body2">
+                    Already have an account? Sign in
                   </Link>
                 </Grid>
               </Grid>
             </form>
           </Paper>
-          <Box mt={8}>
+          <Box mt={5}>
             <Copyright yourWebsite="My Diary App Demo" />
           </Box>
         </Container>
